# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mistral.rs",
    "description": "Fast, flexible model inference.",
    "contact": {
      "name": "Eric Buehler"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.9.3"
  },
  "paths": {
...
    "/v1/agent/approvals/{approval_id}": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "resolve_agent_approval",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "description": "Approval ID from the approval-required SSE event",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Decision applied or queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decision payload"
          },
          "404": {
            "description": "Unknown approval ID"
          },
          "413": {
            "description": "Decision payload is too large"
          },
          "415": {
            "description": "Unsupported content type"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalDecision": {
        "type": "string",
        "enum": [
          "approve",
          "deny"
        ]
      },
      "ApprovalDecisionRequest": {
        "type": "object",
        "description": "Decision payload for a pending agentic tool approval.",
        "required": [
          "decision"
        ],
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional note passed back to the model on denial."
          },
          "remember_for_session": {
            "type": "boolean",
            "description": "Auto-approve all later tool calls in the same session."
          }
        }
      },
      "ApprovalDecisionResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "\"resolved\" or \"queued\"."
          }
        }
      },
...
    }
  },
...
}